Support cETH
This patch starts to add support for cETH to both the spec and the code. This is necessary since cETH will be a common collateral token.

We also start to build the sweep function.

We also start to implement `inMigration`.

We also slim out Uniswap libraries.

We also start to fix token approvals and remove the "enumerated collateral tokens" concept.