Add Even More Tests
As stated, just adds a variety of more tests, focusing on getting us to a full suite of tests. Honestly, most of these are less likely to be risks since a) in the worst case you can't repay the flash loan so its unlikely you're at risk, and b) they are mostly unrelated to operator security (e.g. someone using the operator in a malicious way). But it's none-the-less worth collecting as many tests as we can build.