Add Goerli Types
This patch adds Goerli types, attempting to keep type-safety in a way that's probably deeper than needed, but generally is consistent and working fairly well at this point. This is working against our newest deployments and would be a good deploy target to test in the app.

Note: we still don't actually do any migration, that's coming next.