Start to Paint a Little
Here we start to paint the app just a little to make it 8% more bearable to look at. We steal copious amounts of styles (and excess styles) from the Webb3 code, which we'll trim down once we pull in the styles we really need. Biggest open issues are getting the asset rows formatted better, making sure things are spaced correctly, adding the position summary boxes, and figuring out how the iframe should flow within the overall embedded page.