Handle Ext Messages and Set Theme
This patch adds a basic set-up for handling inbound messages that aren't RPC calls in an extension. For instance, the Webb3 app will now send a message to an extension when the theme changes, so the extension can sync on light or dark themes. This will also be a basis for proactively telling an extension that the user's account or network has changed, so we don't need to wait for a poll in the extension to learn that information.