Fix Type Issue
This patch fixes a type issue that seems to arise from tsc not properly knowing if a key is defined in a map, which could return undefined but it thinks it would always return `Promise<any>` (that should be `Promise<any> | undefined`). Anyway, we switch from `if (cache[x])` to `if (x in cache)`, which should be identical for our purposes and later instead of `cache[x] = null as any` instead we use `delete cache[x]`. I'm not 100% positive this will work, but it _should_.