Easier Playground Deploy
This patch makes playground deployments easier, by automatically picking up the migrator address. We write the address when we deploy to a "playground.env.local" file and then we start vite in `--mode playground`, which picks up the address from that file. This lets us use that file to start the playground, but doesn't accidentally pick up that address when running in dev/prod (which otherwise now comes from `.env`). Overall, this is meant as a start to simplifying the development flows for testing/building extensions.