Add Mainnet Playground
This adds a mainnet playground that doesn't override what's already on mainnet and just uses it. I tried to unify these three scripts, but it ended up being a mess of weird minor deps inconsistencies. I might try to consolidate the `playground.*.sh` scripts in the future, and make the deployer not the primary user (so its nonce starts at 1, not 9). This should, nonetheless, be good to go as is.