Test that base asset collateral are migrated properly (#53)
In this test, we migrate only the collateral of type base asset (USDC) while taking out a flash loan. We expect the base asset collateral to be properly migrated to Compound III, with some amount taken off to repay the flash loan interest. We expect the migrator to not retain any of the user's base asset.