Don't borrow from Comet if migrator has enough base tokens (#51)
There is a weird edge-case where a migration will revert if the migrator has a base token balance greater than `flashAmountWithFee`. This would only happen if someone sends/donates funds to the migrator. To handle this case, we no longer borrow from Comet if there is already a large enough base token balance present in the migrator to pay back the loan.