Clean up Playground
This patch consolidates the Playground scripts to make them easier to manage, specifically merging the bash scripts into one and using env vars to differentiate "re-deploy" versus "use mainnet". We keep the V1 and V2 Playground scripts separate, since they end up _not_ sharing a lot of overlapping types, due to weird ERC20 support (to cover USDT in V2).

Finally, we also make sure the account nonce is always 10, to make it clear what the nonce value should be when running transactions, making it less likely to end up in a weird "pending" state if the nonce is incorrect.