Use exact amounts instead of `balanceOf` to move collateral around (#55)
Bringing back the implementation in #46 to use exact amounts for moving around collateral instead of `token.balanceOf(address(this))`.

This is needed to patch up an edge case where a migration would fail if the same asset is migrated as both collateral and borrow. This can only happen for the base asset (USDC). 

**Example edge case**
User wants to migrate 100 USDC worth of cUSDC and 100 USDC of Aave v2 borrows using a flash amount of 100 USDC. The order of operations is as follows:
1. Flash 100 USDC to migrator.
2. Redeem 100 USDC worth of cUSDC from Compound V2.
3. Supply `USDC.balanceOf(address(this))` (200) of USDC to Compound III.
4. Repay 100 USDC of Aave v2 borrows.
5. Revert because there is 0 USDC left in the migrator to repay the borrow (the 100 USDC flash amount was supplied to Compound III in step 3).