Migrator (#36)
* Migrator push

* Add CometState Message Handler

This patch adds a simple CometState message handler.

* Memoize RPC on

* Bump Comet Extension

* Rework, minimally

* Bump version of comet-extensin

* Add Transaction Tracker

The transaction tracker keeps track of the receipts for transactions. Curerntly we're just using this value as a "when status changes, reload page data" but we could display a modal, prevent other transactions, or show logs, etc, based on this data that is now tracked.

* Poll 30s instead of 10s, since we reload after trxs

* Disable buttons while we have pending transactions

* Checkpoint before I lose more work in a rebase =[

* Add pending states and modals

* Bump comet-extension version

* Properly account for V3 existing borrow balance

* get playground to v1 migrator

* Playground V2 and other Enhancements (#52)

This patch, firstly, adds a `yarn playground:v1` and `yarn playground:v2`. This allows us to simply run a playground for either version of the contract.

Secondly, we change our mnemonic phrase to something a little more random, so we don't fight against a very commonly used private key. That said, this new key should be considered 100% unsecure, but at least we don't see random trxs committed by it every few days on mainnet.

Thirdly, we set `strictPort=true` to hopefully prevent running a playground accidentally twice.

* Remove console.log

* Update to use liquidationCollateralFactor not liquidationFactor

* Add Mainnet Playground (#54)

This adds a mainnet playground that doesn't override what's already on mainnet and just uses it. I tried to unify these three scripts, but it ended up being a mess of weird minor deps inconsistencies. I might try to consolidate the `playground.*.sh` scripts in the future, and make the deployer not the primary user (so its nonce starts at 1, not 9). This should, nonetheless, be good to go as is.

* Use ethers multicall (#56)

* Update requests to ethers-multicall

* Clean up Playground

This patch consolidates the Playground scripts to make them easier to manage, specifically merging the bash scripts into one and using env vars to differentiate "re-deploy" versus "use mainnet". We keep the V1 and V2 Playground scripts separate, since they end up _not_ sharing a lot of overlapping types, due to weird ERC20 support (to cover USDT in V2).

Finally, we also make sure the account nonce is always 10, to make it clear what the nonce value should be when running transactions, making it less likely to end up in a weird "pending" state if the nonce is incorrect.

Co-authored-by: Geoffrey Hayes <hayesgm@gmail.com>

Co-authored-by: Geoffrey Hayes <hayesgm@gmail.com>