Der Great Hack in the Sky
This patch is to fix an issue where rollupjs is (rightfully?) missing deps when building Uniswap libraries, as they seem to misreport their requirements. This is only manifesting itself from rollup, not from esbuild, so it's becoming a pain to figure out exactly why and how to address it. To get things going, we are forcing the built/optimizied/minified library to rely on other deps, by hacking into the standard output and looking for `a.BigInt()` and `l.BigInt()` and replacing them with `A.BigInt()`. This is similar to using externals and globals in Rollupjs, but that's _also_ not working very well for these libraries. This is intended to be removed when we can deeper address the issue.